requirement.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
transformers==4.35.2
torch==2.1.0
spacy==3.7.2
scikit-learn==1.3.2
flask==3.0.3
flask-sqlalchemy==3.1.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
schedule==1.2.0